+++ lega/conf/defaults.ini
@@ -19,12 +19,12 @@ mode = 2750
 ###########################
 location = /ega/vault
 mode = 2750
-driver = lega.vault.FileStorage
+driver = FileStorage
 
 ###########################
 # Backed by S3
 ###########################
-# driver = lega.vault.S3Storage
+# driver = S3Storage
 url = http://ega_vault_s3:9000
 region = lega
 bucket = lega
