--- tests/openpgp_data.py
@@ -1,108 +1,0 @@
-PGP_PASSPHRASE = b'crazywow'
-PGP_NAME = 'PyTest'
-PGP_COMMENT = 'Fake Key for PyTest'
-PGP_EMAIL = 'fake@pytest.ega'
-
-PGP_PUBKEY = '''\
------BEGIN PGP PUBLIC KEY BLOCK-----
-
-mQINBFq//YgBEACw2R0ZfzYFaO0A9w3A0IP3H0fed65DYitcCQa2D3iuDfIHNmEl
-sjEExeisumpzhTSko8hpCUx6k9c4cnaW+iexCaIKZVanjpR0HSuELFDXnEvzjwys
-SLI8kyaTtijlva9NMz0iPXY50ZPsMxfsadzIowui3MHf/zE4OiHFD5ioLoW42Ums
-SS9sS2HapyBsiBdf8vbCO3hRZ+KiDMhHz1rfoGAFQ7qRdyg8jDZG6Jp+Wfi9W+QJ
-zO1IdKmihlbIXtRY6pN9g1sRK1FuQ7ckqkjpJDs5KO8odSLfw4JcYbdi+/GdPRoG
-+v3R89omK54aDUa4APctOzHD+Ir7WSUKqIBAuzoFCCYOlbzwNDCwh8gzXpwYsZ5n
-xnFpqJQqZ9L4qoSe33agy2pBB9iGcWGr7qGCKxByxoCPXA8kwKfADFOnqMOxnufD
-OveDnPgvlHHQhcaE8SkH59N0yZqZPxvLvWPt96aYZfFT06yuFFrFdHmJ8TJM7/og
-T72Urh+TkvMc6vlBqrECSkkWsO1sQAbGEGT5BtqQT1NYWemQoo107KiahqUS4YfA
-SMuoQAufNVHeIptA7Oxz2as91c1XPRpHORUKy5l1zMmRXgdgHpUnwIRtIztS7ZN7
-uZyvsyrPdMIowIfGi8gkRDpwHICREnIZU+HLV9HWJ6tkEMoDII3FdNXe3QARAQAB
-tC5QeVRlc3QgKEZha2UgS2V5IGZvciBQeVRlc3QpIDxmYWtlQHB5dGVzdC5lZ2E+
-iQJOBBMBCAA4FiEEBr3SCLstcJrDVCtX5Sw8q8eIhpUFAlq//YgCGy8FCwkIBwIG
-FQoJCAsCBBYCAwECHgECF4AACgkQ5Sw8q8eIhpW+Iw/+J3cYWYmdM/cRoFkjF3nD
-1YWPHkNtViVP30KWtQE9SEdmVmchbIXGWvfWvNxREHYr2c2kVZqgcNYbfYQ/X4EN
-t5jFrmRg9Ab1gcVfNesDZ3fYDGi070iTLG3XRYqnd8ljX80hNmem9hx45DOWjIS6
-Nl4niOCcEDtMR1WQijWO3bJydKyTvqD/muAfcOvskAz2NNYUHcARnb7SOSuMx7kd
-+LWG8i91XsQklXipOCHMnEMz04ULkHeoOj9zAJ4ekvmZaNW5/9rB7PuviBwwed6e
-JIDqI9wbQKqQwRrVfa7fFK5sxYZCApDKr4uljc7W2QP1Yhtg6Sa2Npwy+qoxq7DX
-rHGLnOVTdOQhuk0zWqXrgW/CrHJOTdqeFdtfQTqyp3oNFvIphyjm0DMUQtyhRMoW
-dEk7xyiAWAme5smGJ0QEVbQurmkgVoWT48TNshYQn3sUKe8Hr7zr2IqwRgszfpwz
-64UPOoidtFQC9xsB2q70wEoqLo0Hn88fwXBzLji++UwEFIHOd7OzMAJuJ9kph5mV
-dk5VOhL8U0PBtd0m8DaMaJVOUtm7tmiM3lWsWAKaSR+5FEOi51zsgSj2xMgLyb/T
-mNXeCLCe3XM4dv90EHthNmQ3fm2TJT6g7eX353xXNb6kJT3GAdd78Gua0LxIq96v
-XOtbHLz8/MwCVsep+ABHVTs=
-=3M+b
------END PGP PUBLIC KEY BLOCK-----'''
-PGP_PRIVKEY = '''\
------BEGIN PGP PRIVATE KEY BLOCK-----
-
-lQdGBFq//YgBEACw2R0ZfzYFaO0A9w3A0IP3H0fed65DYitcCQa2D3iuDfIHNmEl
-sjEExeisumpzhTSko8hpCUx6k9c4cnaW+iexCaIKZVanjpR0HSuELFDXnEvzjwys
-SLI8kyaTtijlva9NMz0iPXY50ZPsMxfsadzIowui3MHf/zE4OiHFD5ioLoW42Ums
-SS9sS2HapyBsiBdf8vbCO3hRZ+KiDMhHz1rfoGAFQ7qRdyg8jDZG6Jp+Wfi9W+QJ
-zO1IdKmihlbIXtRY6pN9g1sRK1FuQ7ckqkjpJDs5KO8odSLfw4JcYbdi+/GdPRoG
-+v3R89omK54aDUa4APctOzHD+Ir7WSUKqIBAuzoFCCYOlbzwNDCwh8gzXpwYsZ5n
-xnFpqJQqZ9L4qoSe33agy2pBB9iGcWGr7qGCKxByxoCPXA8kwKfADFOnqMOxnufD
-OveDnPgvlHHQhcaE8SkH59N0yZqZPxvLvWPt96aYZfFT06yuFFrFdHmJ8TJM7/og
-T72Urh+TkvMc6vlBqrECSkkWsO1sQAbGEGT5BtqQT1NYWemQoo107KiahqUS4YfA
-SMuoQAufNVHeIptA7Oxz2as91c1XPRpHORUKy5l1zMmRXgdgHpUnwIRtIztS7ZN7
-uZyvsyrPdMIowIfGi8gkRDpwHICREnIZU+HLV9HWJ6tkEMoDII3FdNXe3QARAQAB
-/gcDAv6oF0chsGZS5F5q7XhXakRN+F8AoAhdoVSU227vis1rbb+fSF+Nt9V6upu3
-JR4l3s8HVQCSWBNsPfyQvuPLj8Fs0wHgsa3atTJWSaTIhlcGFLMc6LynUsUyV9xW
-U9X619jBCB+pu3I3xHev9lrARfz3Z2viHE6bu2djSX9zHTuNAagxHLx7tUY7LyWK
-JhOF5a/wb3GxrHAsmxWrA7gh11lWFzfMdDBcaTiBY8jcFu+tmIwc1482oAOpbOVU
-Lh5vKQJH7EDwU8jkla6RkC/JO/ZZ9CUrJkjlHbvoMFuBERMYBPD5CIM/hd+ADiwo
-zhKvFcX/MXhF7RmckMxJOKsTPWDQpdFHWjUTefr00SLVmiNbgl/fVJZ/by+/ZSjd
-3tfSuvgyTphpumfQRi3yv0UL+POiVInyJ14qcIAOAVY9/svuIZ2T9WIbsA6OIzU4
-bPOIeIlSOktuLlXCwUPW7J6/WJ9yxCAhrkfFiVFgYm5E/W2WvGSvKC2S2wW/xnPQ
-wyf3bSxHqyNKf1Sd6XqcyC7LUw+4aSeTue4WyVcYZEG1D2haG3RNRRKdYGXRduGP
-3cPQSQjL5feJLcoTPus8vg0P71o7SeSwhyWqTaTVQ4dLO8TDdKc0XXR2ERANz2re
-hNg+ziqAt0vA7BXC/dEMtlGj8N9eg71LUp/uTyT1W3t3FBe8vaXdbmCb/mD5hWWm
-i9VUinz60iW2bRB2SvzeQFfgIThT8XyfJCkg/d0PGAv89ll1a3zX04ajvApG+6hC
-Bo7xG2p4OQ091EUaPbqdtJUbsObAhXIam8y/BBzKzCJAXjZ1lfkOFvOdW4dKCNsT
-68OzOGzIN9bn+8sRgYB9GWhtmnXYwtsbv43pNG6wEMiaSgn/B1pdgdu6GJC4BYzk
-j/+Dpq2I9As76cFV6m5kyMm1D9/GVnR/z0V++aoMM3kPH0ZAEFJun7iobW6k4u5z
-c29MfcbOzBM9IMdClC0qHptm6IHAjdQfQeCG6x9U1zLIf/fQx5OsiAUO21LgJyjF
-s/i/9dbtwpqybPISO+HQxznmW09cWt2m1ReDRbBLDpIGR9VaSa8PGl8dVxmutTwq
-QVXi128+wMuBpKQLazDqjSWGlr5DcmUrHV5PzyxoWtFPyJrsSr5g6UsueS+qOas1
-H71dvlCwlEc61q3LKbVFXsTN15qZs7786jgSLp5oyXCLkE6gKARs4zWHRE5zNiBp
-t/kCbo/yuL+vpUbW77rqknHx/Sc53pOb8tUAycUEZVJDfRVx15P9e1ZZuVq6RnIg
-8/f0YabL7HFCkg/zh3zZPYoparOXl1eZVDlXZQa2q4H7Z5218YAY5GwyYDti2g7P
-ZAhZs1dG0pdnuukEg23XmmgTh4JwKi2amIMZ6cS9vosHQKF3T1lVpp/JQoMEqa8s
-M1n3qabVghV/QMrtwKg9Pau2AwOc92XEtx+TpBtdZd66Rk7m63Q90JIYGKk1rZ2g
-SMOyju3OrVF8JcpULTp475LBHl2Mpmy8I2m1CjDA65b7cV0EXg3I+v2e/Yswsdwu
-dnD88uIMGxMG0ziVfdPH5Y5XCnywWANPXxP71GJhBqDVljOlg+BLym158u0ueqwH
-O2z7Idny3KVKBR2BpP3Piz0DlA3gXftyQpkBv8JxMLfbd55wOT5DWZUzDcqaccbh
-ofv/CUameNuXHwG5e4Oe1um38aJ5yW7dYVdMN0bdazXSy2O1ADqtJQHZYRBVYVzX
-xFkKgYWgKpVShOBSe3yWD3GpaGwU/tE65SS5Yz1eXDmNM2i0+RvuT4K0LlB5VGVz
-dCAoRmFrZSBLZXkgZm9yIFB5VGVzdCkgPGZha2VAcHl0ZXN0LmVnYT6JAk4EEwEI
-ADgWIQQGvdIIuy1wmsNUK1flLDyrx4iGlQUCWr/9iAIbLwULCQgHAgYVCgkICwIE
-FgIDAQIeAQIXgAAKCRDlLDyrx4iGlb4jD/4ndxhZiZ0z9xGgWSMXecPVhY8eQ21W
-JU/fQpa1AT1IR2ZWZyFshcZa99a83FEQdivZzaRVmqBw1ht9hD9fgQ23mMWuZGD0
-BvWBxV816wNnd9gMaLTvSJMsbddFiqd3yWNfzSE2Z6b2HHjkM5aMhLo2XieI4JwQ
-O0xHVZCKNY7dsnJ0rJO+oP+a4B9w6+yQDPY01hQdwBGdvtI5K4zHuR34tYbyL3Ve
-xCSVeKk4IcycQzPThQuQd6g6P3MAnh6S+Zlo1bn/2sHs+6+IHDB53p4kgOoj3BtA
-qpDBGtV9rt8UrmzFhkICkMqvi6WNztbZA/ViG2DpJrY2nDL6qjGrsNescYuc5VN0
-5CG6TTNapeuBb8Ksck5N2p4V219BOrKneg0W8imHKObQMxRC3KFEyhZ0STvHKIBY
-CZ7myYYnRARVtC6uaSBWhZPjxM2yFhCfexQp7wevvOvYirBGCzN+nDPrhQ86iJ20
-VAL3GwHarvTASioujQefzx/BcHMuOL75TAQUgc53s7MwAm4n2SmHmZV2TlU6EvxT
-Q8G13SbwNoxolU5S2bu2aIzeVaxYAppJH7kUQ6LnXOyBKPbEyAvJv9OY1d4IsJ7d
-czh2/3QQe2E2ZDd+bZMlPqDt5ffnfFc1vqQlPcYB13vwa5rQvEir3q9c61scvPz8
-zAJWx6n4AEdVOw==
-=l9D8
------END PGP PRIVATE KEY BLOCK-----'''
-
-PGP_PUBKEY_BIN = '99020d045abffd88011000b0d91d197f360568ed00f70dc0d083f71f47de77ae43622b5c0906b60f78ae0df207366125b23104c5e8acba6a738534a4a3c869094c7a93d738727696fa27b109a20a6556a78e94741d2b842c50d79c4bf38f0cac48b23c932693b628e5bdaf4d333d223d7639d193ec3317ec69dcc8a30ba2dcc1dfff31383a21c50f98a82e85b8d949ac492f6c4b61daa7206c88175ff2f6c23b785167e2a20cc847cf5adfa0600543ba9177283c8c3646e89a7e59f8bd5be409cced4874a9a28656c85ed458ea937d835b112b516e43b724aa48e9243b3928ef287522dfc3825c61b762fbf19d3d1a06fafdd1f3da262b9e1a0d46b800f72d3b31c3f88afb59250aa88040bb3a0508260e95bcf03430b087c8335e9c18b19e67c67169a8942a67d2f8aa849edf76a0cb6a4107d8867161abeea1822b1072c6808f5c0f24c0a7c00c53a7a8c3b19ee7c33af7839cf82f9471d085c684f12907e7d374c99a993f1bcbbd63edf7a69865f153d3acae145ac5747989f1324ceffa204fbd94ae1f9392f31ceaf941aab1024a4916b0ed6c4006c61064f906da904f535859e990a28d74eca89a86a512e187c048cba8400b9f3551de229b40ecec73d9ab3dd5cd573d1a4739150acb9975ccc9915e07601e9527c0846d233b52ed937bb99cafb32acf74c228c087c68bc824443a701c809112721953e1cb57d1d627ab6410ca03208dc574d5dedd0011010001b42e507954657374202846616b65204b657920666f722050795465737429203c66616b65407079746573742e6567613e89024e04130108003816210406bdd208bb2d709ac3542b57e52c3cabc788869505025abffd88021b2f050b0908070206150a09080b020416020301021e01021780000a0910e52c3cabc7888695be230ffe27771859899d33f711a059231779c3d5858f1e436d56254fdf4296b5013d4847665667216c85c65af7d6bcdc5110762bd9cda4559aa070d61b7d843f5f810db798c5ae6460f406f581c55f35eb036777d80c68b4ef48932c6dd7458aa777c9635fcd213667a6f61c78e433968c84ba365e2788e09c103b4c4755908a358eddb27274ac93bea0ff9ae01f70ebec900cf634d6141dc0119dbed2392b8cc7b91df8b586f22f755ec4249578a93821cc9c4333d3850b9077a83a3f73009e1e92f99968d5b9ffdac1ecfbaf881c3079de9e2480ea23dc1b40aa90c11ad57daedf14ae6cc586420290caaf8ba58dced6d903f5621b60e926b6369c32faaa31abb0d7ac718b9ce55374e421ba4d335aa5eb816fc2ac724e4dda9e15db5f413ab2a77a0d16f2298728e6d0331442dca144ca1674493bc7288058099ee6c98627440455b42eae6920568593e3c4cdb216109f7b1429ef07afbcebd88ab0460b337e9c33eb850f3a889db45402f71b01daaef4c04a2a2e8d079fcf1fc170732e38bef94c041481ce77b3b330026e27d929879995764e553a12fc5343c1b5dd26f0368c68954e52d9bbb6688cde55ac58029a491fb91443a2e75cec8128f6c4c80bc9bfd398d5de08b09edd733876ff74107b613664377e6d93253ea0ede5f7e77c5735bea4253dc601d77bf06b9ad0bc48abdeaf5ceb5b1cbcfcfccc0256c7a9f80047553b'
-PGP_PRIVKEY_BIN = '950746045abffd88011000b0d91d197f360568ed00f70dc0d083f71f47de77ae43622b5c0906b60f78ae0df207366125b23104c5e8acba6a738534a4a3c869094c7a93d738727696fa27b109a20a6556a78e94741d2b842c50d79c4bf38f0cac48b23c932693b628e5bdaf4d333d223d7639d193ec3317ec69dcc8a30ba2dcc1dfff31383a21c50f98a82e85b8d949ac492f6c4b61daa7206c88175ff2f6c23b785167e2a20cc847cf5adfa0600543ba9177283c8c3646e89a7e59f8bd5be409cced4874a9a28656c85ed458ea937d835b112b516e43b724aa48e9243b3928ef287522dfc3825c61b762fbf19d3d1a06fafdd1f3da262b9e1a0d46b800f72d3b31c3f88afb59250aa88040bb3a0508260e95bcf03430b087c8335e9c18b19e67c67169a8942a67d2f8aa849edf76a0cb6a4107d8867161abeea1822b1072c6808f5c0f24c0a7c00c53a7a8c3b19ee7c33af7839cf82f9471d085c684f12907e7d374c99a993f1bcbbd63edf7a69865f153d3acae145ac5747989f1324ceffa204fbd94ae1f9392f31ceaf941aab1024a4916b0ed6c4006c61064f906da904f535859e990a28d74eca89a86a512e187c048cba8400b9f3551de229b40ecec73d9ab3dd5cd573d1a4739150acb9975ccc9915e07601e9527c0846d233b52ed937bb99cafb32acf74c228c087c68bc824443a701c809112721953e1cb57d1d627ab6410ca03208dc574d5dedd0011010001fe070302f2ff7b8bc75b693ae68f70c299b1ce727bdf10903b75d063bea62dee42ef3dd5781ae03bf67f6b363877f38000352435d1aa02cff7669e9291522fe71b199f3e2a621ebfdcbdbb5388042ef0b8099e2930d872740d326e275c0de20541076198b7d31b6fdf9a02ad874dbae26dd6dff6f434812048884709fc6d4af3786f10ede4f53580c3a83382bd6b1aa2f5402e6c0d649a480a763b0213394737feb689cf5da1ff804d01dbbc1d422227e8e361ce6253b2089ca2e19712212b2ac98d0ee15b2008e7666fe771b3a0ca8945e84323c9ea84de7ea7d2c90d9652a9726a07d9373b293e0957ccf694ca72d0160de8635dc8d7020331868c86017a6b2e445c5fc9624d7005e793bb4ddcb16c55b1d30d13f80df933b4f5e8597eda542b570c65c63de026bfa23938afdacdd32092023732250fd3ab36b1052d68e2b8d26970b1ce05044c715ddac82c25a6f414c82f1f230f1978012300ebdd5efeec7048100d4d7857d6fe7f699b55b2abcbeb19b542d8f5739f19882c36f07207fd9ac4311adaf4b2f74351f8d4b9854b3a7bf8d807f1d65550d5c09688a74a657e7abf8cc60e64a1dab38481b3f336eb1ee9604e89fb95b7f8d6bace21566f4aa397f1bbab80b57d01d1a780d1e506cefb4389213cddbc31eb41a6a6254f8c5c9e8b44c542414a363b2c683986045c2a4232f41135d58b8c85f16bc0bd4898071137998917ef56a052e416fb74f9ae25d9e06e10779e06a044ba784e97d2a889b3b50955ae2245a04bd14f1bc363b8a6bfd76b0ab487043555c220fc776a2decb967a0c7c9f274d36740e4b125d002a38746461c78b518afb9ccb6e1fa9640608ecd2df60667603ce0d48f2dca147f6be30aa6fe4d2d52409aa957625cce10206b97f7a38bc6b8fe00e70d1d0b0bdabd6facb6723abd326dd479c26e6080a488cb7034325d3ac20994edd86761ce200efbc2fa4dc9c9421ce90657a805fd5f56b0254892aebed07e5c57055a4a2a7f2bfc302cc1f35a450ba4086e0620f09c3d95c7b83486fb8b5ca6632e25d84a513c171fc15e229dd91fc6b46cda01d33bdc11bd675da26af653cf2ddfa5a410810f05794753778fe6ae430189c50243ca6185ff14c72bc82ba211377c03c29be40a2e81b0bf763c45271fb50c9000a6cc11acb693905a3e17e65ca7b8c1aac49b336d74d9d443126a2f1a96d416cf6e683f2ca47ee8809f831e0364965504595157fb4560521cbf03389bcd1f500fe77e4f54830aa281dcc6b9f00bb48ec627b83e669b0e5e391a32de2725e1b79f4664b5774192298b30fc8c5c0a6ec1e1f64ce6ee0e84804a0b5b732447f14a86cbe3699915a42ff12ec9c3305b10f91e39e792038af8286009ed9febf9c91d1dd162c57bac12390e3d77c45e332b1d430a94997116d637b120f415a3f15d1c9dcd6f91a34eb0c995c98fd8b7ea1c82e457040a2cdc102e16af1dfd9cebeaffd4901b42b680c62893c075601377bf27afc3bee0ba3a048fccedd21b32eb369c5afdc7bdbd0f7adba23b389911284c0e38e4d3ba80891b11e870da85b38d65939998fdeaf3251307f2d0e504b3a37ef15f7244f5ef997d2b4cf10ee688ac78c44eff8dfe1968d1b8d702e5e775f97d01fa7857c890113abd200c672f0e4c480201092513d9c9020454caa6e9ad1b781fb91b6df112b8b8eb415e7ef49a1cbad70e9a1bd7f3be3d5aacbd7144364675b3fd307c25bebe3fc7d5627e03b4ca8737e65215aaff5934c29ae663919e6823cd2d54308a5a31fceee5630bf7da97db10e6a12c89e274f14549c64cf551f8d9a8d221f60f6501fd7d393306c7887042c4fc22ba622af3c3dd30059fde228b17b513148401065fa4398bbc2dd77788b42e507954657374202846616b65204b657920666f722050795465737429203c66616b65407079746573742e6567613e89024e04130108003816210406bdd208bb2d709ac3542b57e52c3cabc788869505025abffd88021b2f050b0908070206150a09080b020416020301021e01021780000a0910e52c3cabc7888695be230ffe27771859899d33f711a059231779c3d5858f1e436d56254fdf4296b5013d4847665667216c85c65af7d6bcdc5110762bd9cda4559aa070d61b7d843f5f810db798c5ae6460f406f581c55f35eb036777d80c68b4ef48932c6dd7458aa777c9635fcd213667a6f61c78e433968c84ba365e2788e09c103b4c4755908a358eddb27274ac93bea0ff9ae01f70ebec900cf634d6141dc0119dbed2392b8cc7b91df8b586f22f755ec4249578a93821cc9c4333d3850b9077a83a3f73009e1e92f99968d5b9ffdac1ecfbaf881c3079de9e2480ea23dc1b40aa90c11ad57daedf14ae6cc586420290caaf8ba58dced6d903f5621b60e926b6369c32faaa31abb0d7ac718b9ce55374e421ba4d335aa5eb816fc2ac724e4dda9e15db5f413ab2a77a0d16f2298728e6d0331442dca144ca1674493bc7288058099ee6c98627440455b42eae6920568593e3c4cdb216109f7b1429ef07afbcebd88ab0460b337e9c33eb850f3a889db45402f71b01daaef4c04a2a2e8d079fcf1fc170732e38bef94c041481ce77b3b330026e27d929879995764e553a12fc5343c1b5dd26f0368c68954e52d9bbb6688cde55ac58029a491fb91443a2e75cec8128f6c4c80bc9bfd398d5de08b09edd733876ff74107b613664377e6d93253ea0ede5f7e77c5735bea4253dc601d77bf06b9ad0bc48abdeaf5ceb5b1cbcfcfccc0256c7a9f80047553b'
-
-ORG_FILE = b'Hello PyTest\n'
-ENC_FILE = '85020C03E52C3CABC788869501100082E9DF4C2A9618F1FCC93BED449011D0DEF17711EB39811857F4968D9A2CAC8A07EA76A267C4B9CBB8CD9660AC8BE22CF8A57CD037EB0D21A351BC884E5D98AF85FA45F81DC96A21B153610D90B4507AA54A9AE909548C7ED9526C0C4B0D2431E6A05E8E149BD00BE350F3683C0977D376269B720E08A6AF7A03D69F320B55CB2C43F3D4FCA3F27C7490AD3840F02DF57A7289320DE45AC6C7489C93FB9C56BD05940209BCE4BE2E969A09BEF6543E7D30BA852BE4396FC48E6741076A809D5E5563DC2585E0AE6C6D0E5D35668F327AF8F115AB97370C7859E9E39BF86281FA03080410C96E25BE20CF81A91291C149704395F635AEF9DE324A207099C4B1CA8360EA2B65CC6BCEEEAD3095EE15A7D71BB5C005951B15D6A41305F0124CC753B3C6878FB10F9DC49261394B0B3999B32322C66522EC44260AA0BDD2F8E62CC6CAE24908EACA3C181EAA643E03EBA37B03F76C3E0023BE37A09DED37E85EDD6512CF73548EFDFC81199CE9200CB88C1A2FB401E939025E6449C70455F415CE05F585C8ECFC5A7FD90D52756CBD6D220755F3B1E509052A1CCF0589D0111EC37BF1FECCA5D91891FB216F921655C49CA29BD73C14FBE51E3608F124288974CCB1D4306E83E2CC659D5DA9AF60E6E7A0CA142240CF664A6C3877D15CBA6F1EF57B3D4A9F51A47141D59897BB85ABCF179E585077D44DF20E7AFA0415D61C0BAB0FD25201384278D02EA162773B43DFEC4B56C223E1477964AB0222A3BF108F703D3DB62BBB534654349DF3A51E35856E48C116E280DCFC193B8E1E28A4A349B9CC7847BBC0C16AE817D3333AFDFE027938328232EB'
-
-ENCRYPTION_KEY_TYPE = 'RSA'
-ENCRYPTION_KEY_BITSIZE = 4096
-KEY_ID = 'E52C3CABC7888695'
-KEY_CREATED = '2018-03-31'
-SESSION_KEY = '92336327AE4ADCD7D0989B9F216CBACCA63AAB9E64DC4C95B74D3CEF130C3764'
-# algo is 9
-
-PGP_PRIVKEY_MATERIAL = {'type': 'rsa', 'public': {'n': 'b0d91d197f360568ed00f70dc0d083f71f47de77ae43622b5c0906b60f78ae0df207366125b23104c5e8acba6a738534a4a3c869094c7a93d738727696fa27b109a20a6556a78e94741d2b842c50d79c4bf38f0cac48b23c932693b628e5bdaf4d333d223d7639d193ec3317ec69dcc8a30ba2dcc1dfff31383a21c50f98a82e85b8d949ac492f6c4b61daa7206c88175ff2f6c23b785167e2a20cc847cf5adfa0600543ba9177283c8c3646e89a7e59f8bd5be409cced4874a9a28656c85ed458ea937d835b112b516e43b724aa48e9243b3928ef287522dfc3825c61b762fbf19d3d1a06fafdd1f3da262b9e1a0d46b800f72d3b31c3f88afb59250aa88040bb3a0508260e95bcf03430b087c8335e9c18b19e67c67169a8942a67d2f8aa849edf76a0cb6a4107d8867161abeea1822b1072c6808f5c0f24c0a7c00c53a7a8c3b19ee7c33af7839cf82f9471d085c684f12907e7d374c99a993f1bcbbd63edf7a69865f153d3acae145ac5747989f1324ceffa204fbd94ae1f9392f31ceaf941aab1024a4916b0ed6c4006c61064f906da904f535859e990a28d74eca89a86a512e187c048cba8400b9f3551de229b40ecec73d9ab3dd5cd573d1a4739150acb9975ccc9915e07601e9527c0846d233b52ed937bb99cafb32acf74c228c087c68bc824443a701c809112721953e1cb57d1d627ab6410ca03208dc574d5dedd', 'e': '010001'}, 'private': {'d': '09361c8429625c089c5456749eaf030f523a33d8aa2e192a56d64ddac4a1084ddd25ff8632b35e9c17f3e2855389f602fe8807bbb82fdb304bff9a6f5a0b5319c4a843a030d1deefc016e1df4b50a053b98808e82e7ec5dbe3b298c888f0f87eed9ce6345c47b0c372ba94e9b6d71d1877ecf6a6c6d41443d120d95a9d3bef6ca6eb45d41c221ebdd10a5a86d07df40910516f3016eebc2cc15052ad49979517e74fabb931f9fadf2612368f331fe846119a495e4372c0b196095ced60394fbe81ed76a66265bcab49329eafd17c2b7a8b4d7e10c758e535b3ba8855d2ad7d92cfc1b27ef7cf4727bd001c9e07d39063f72dc3b22a83bbd2fc683da5dbdc2d7be62a04bf8a09217541237b15a4064884fd7f9311ecff029fca50f3238f0a1b8297ecb78f437ac8922c67363520797ca7d74662b176c6bf1837fcdeeedd3a0f880c88906e5579289658ae2a92e1793cc9c67af7d4e5e83e5403b0a00f08fd53e49b254cc2d74105059567773d78f3707eda2e50a927ca7425ef77225d3c4411cf8ae0d47035262d72dc80f51769036b51a5c88672f332621c3e23fbef6d1fc161a4c5a5ea5f80e9725ee1a1b0dbe3756f6fae6a2b219686545456ced6cfcedf98ff55929a6c0419f7de0e3848caca27d38c69f598e19d6bb2c454fdef52a1fa239de652dcd9cbaf411ea77f0f116be3abc806fb2638fa724e76a31b28ddd69ab9', 'p': 'd2f36e2c3fa88ef4f50ff10eb9fe2ca10f794c6f55efd550728f71ff447a157a9382ee2302d469770dabd949400388fc4d172ef81f990648abc8e13079c0f628242cc43a8b732c2033fc14122defa77bbee88c5a16c834643cd7fb34d1c97a4c89b3ada664047a922fe99dad4163606b90e1722e7634872902554151c926c24f1ab73963a77c109db2c2f2d7560c1a310a2f5dcff7290346feea351a30aca0428ef64b52fb12b89dab7991ebcd2a37976f5dd7ff1b8c87348e913fcb4c380cae370142cec96fbd3e87c15119ae4411ce879c88cb75eeed7cfe18030b07ffe1a922ec15c52daebd67166f56268640a82cca9b62f6092eddf76d9d6568093ef1c5', 'q': 'd69d4d50f85a6d850f575a63c4dba72a170b8ea4aff67930e51e6d96989b338cd08faca24725c3e5302ae610755f223f63d401ea8104e4c7a8f6c1b326d628144b810ea166a3dc7fcb506b73100df86ba3762fd5a0830b3b03004925664150d591074427f9766646a9f5c4ead6ab2fc1834f3e2a39433ba41a26ea30d15ec3a1e513f421fb7591097a28642cbfa7ac89abdfcb735e68e60f0fab3a5837b15ec28a64ce0ceab64cca7912864aa18ee0b13a77f95f9197c12572a8c73c18881fa36c92cde4ccb6f17cbaef5797130fbed084c71f7461aff67c2b1d37181597cc5ec879bfa516fe36004f4802af6b2e64227a21c4fdca54665e8bf866de5f9c8239', 'u': 'a6f32dd863edcbca2285e350adfd04d322e2dcbcd52a89cd37d855f9e75af43a4bc0dc3f40a2d140f0c9d64971421a6d743750830694aae1bad1e57728e7cec84d5167d1c4a72edeb11e0172b46ab814bb440502e512668edf9c622ae8015833a9fcd8b84856746f500d760158eaa6b3f21fd5b2d49c994a3f2ffc91df7cd83d11e6812cf389d17a3aa0f2d052729c7fae30876a35bb412efddd33c96cae3c974489eeef684cd4bf48d7a39c6f52dbbc76dbfdf286ce07e4944bbca028b234183f0a409d0bf2b98d80bce47c2866a4c174905a6b3a1ae38c4273972c2b1214fcad6cb1ed63ddb639b3b743e2c72ae5d69c81a737410fd7271a8dae7619c1c5df'}}
