+++ tests/pgp_data.py
@@ -1,0 +1,112 @@
+PGP_PASSPHRASE = 'crazywow'
+PGP_NAME = 'PyTest'
+PGP_COMMENT = 'Fake Key for PyTest'
+PGP_EMAIL = 'fake@pytest.ega'
+
+PGP_PUBKEY = '''\
+-----BEGIN PGP PUBLIC KEY BLOCK-----
+Version: some version
+
+mQINBFq//YgBEACw2R0ZfzYFaO0A9w3A0IP3H0fed65DYitcCQa2D3iuDfIHNmEl
+sjEExeisumpzhTSko8hpCUx6k9c4cnaW+iexCaIKZVanjpR0HSuELFDXnEvzjwys
+SLI8kyaTtijlva9NMz0iPXY50ZPsMxfsadzIowui3MHf/zE4OiHFD5ioLoW42Ums
+SS9sS2HapyBsiBdf8vbCO3hRZ+KiDMhHz1rfoGAFQ7qRdyg8jDZG6Jp+Wfi9W+QJ
+zO1IdKmihlbIXtRY6pN9g1sRK1FuQ7ckqkjpJDs5KO8odSLfw4JcYbdi+/GdPRoG
++v3R89omK54aDUa4APctOzHD+Ir7WSUKqIBAuzoFCCYOlbzwNDCwh8gzXpwYsZ5n
+xnFpqJQqZ9L4qoSe33agy2pBB9iGcWGr7qGCKxByxoCPXA8kwKfADFOnqMOxnufD
+OveDnPgvlHHQhcaE8SkH59N0yZqZPxvLvWPt96aYZfFT06yuFFrFdHmJ8TJM7/og
+T72Urh+TkvMc6vlBqrECSkkWsO1sQAbGEGT5BtqQT1NYWemQoo107KiahqUS4YfA
+SMuoQAufNVHeIptA7Oxz2as91c1XPRpHORUKy5l1zMmRXgdgHpUnwIRtIztS7ZN7
+uZyvsyrPdMIowIfGi8gkRDpwHICREnIZU+HLV9HWJ6tkEMoDII3FdNXe3QARAQAB
+tC5QeVRlc3QgKEZha2UgS2V5IGZvciBQeVRlc3QpIDxmYWtlQHB5dGVzdC5lZ2E+
+iQJOBBMBCAA4FiEEBr3SCLstcJrDVCtX5Sw8q8eIhpUFAlq//YgCGy8FCwkIBwIG
+FQoJCAsCBBYCAwECHgECF4AACgkQ5Sw8q8eIhpW+Iw/+J3cYWYmdM/cRoFkjF3nD
+1YWPHkNtViVP30KWtQE9SEdmVmchbIXGWvfWvNxREHYr2c2kVZqgcNYbfYQ/X4EN
+t5jFrmRg9Ab1gcVfNesDZ3fYDGi070iTLG3XRYqnd8ljX80hNmem9hx45DOWjIS6
+Nl4niOCcEDtMR1WQijWO3bJydKyTvqD/muAfcOvskAz2NNYUHcARnb7SOSuMx7kd
++LWG8i91XsQklXipOCHMnEMz04ULkHeoOj9zAJ4ekvmZaNW5/9rB7PuviBwwed6e
+JIDqI9wbQKqQwRrVfa7fFK5sxYZCApDKr4uljc7W2QP1Yhtg6Sa2Npwy+qoxq7DX
+rHGLnOVTdOQhuk0zWqXrgW/CrHJOTdqeFdtfQTqyp3oNFvIphyjm0DMUQtyhRMoW
+dEk7xyiAWAme5smGJ0QEVbQurmkgVoWT48TNshYQn3sUKe8Hr7zr2IqwRgszfpwz
+64UPOoidtFQC9xsB2q70wEoqLo0Hn88fwXBzLji++UwEFIHOd7OzMAJuJ9kph5mV
+dk5VOhL8U0PBtd0m8DaMaJVOUtm7tmiM3lWsWAKaSR+5FEOi51zsgSj2xMgLyb/T
+mNXeCLCe3XM4dv90EHthNmQ3fm2TJT6g7eX353xXNb6kJT3GAdd78Gua0LxIq96v
+XOtbHLz8/MwCVsep+ABHVTs=
+=3M+b
+-----END PGP PUBLIC KEY BLOCK-----'''
+PGP_PRIVKEY = '''\
+-----BEGIN PGP PRIVATE KEY BLOCK-----
+Version: some version
+
+lQdGBFq//YgBEACw2R0ZfzYFaO0A9w3A0IP3H0fed65DYitcCQa2D3iuDfIHNmEl
+sjEExeisumpzhTSko8hpCUx6k9c4cnaW+iexCaIKZVanjpR0HSuELFDXnEvzjwys
+SLI8kyaTtijlva9NMz0iPXY50ZPsMxfsadzIowui3MHf/zE4OiHFD5ioLoW42Ums
+SS9sS2HapyBsiBdf8vbCO3hRZ+KiDMhHz1rfoGAFQ7qRdyg8jDZG6Jp+Wfi9W+QJ
+zO1IdKmihlbIXtRY6pN9g1sRK1FuQ7ckqkjpJDs5KO8odSLfw4JcYbdi+/GdPRoG
++v3R89omK54aDUa4APctOzHD+Ir7WSUKqIBAuzoFCCYOlbzwNDCwh8gzXpwYsZ5n
+xnFpqJQqZ9L4qoSe33agy2pBB9iGcWGr7qGCKxByxoCPXA8kwKfADFOnqMOxnufD
+OveDnPgvlHHQhcaE8SkH59N0yZqZPxvLvWPt96aYZfFT06yuFFrFdHmJ8TJM7/og
+T72Urh+TkvMc6vlBqrECSkkWsO1sQAbGEGT5BtqQT1NYWemQoo107KiahqUS4YfA
+SMuoQAufNVHeIptA7Oxz2as91c1XPRpHORUKy5l1zMmRXgdgHpUnwIRtIztS7ZN7
+uZyvsyrPdMIowIfGi8gkRDpwHICREnIZU+HLV9HWJ6tkEMoDII3FdNXe3QARAQAB
+/gcDAv6oF0chsGZS5F5q7XhXakRN+F8AoAhdoVSU227vis1rbb+fSF+Nt9V6upu3
+JR4l3s8HVQCSWBNsPfyQvuPLj8Fs0wHgsa3atTJWSaTIhlcGFLMc6LynUsUyV9xW
+U9X619jBCB+pu3I3xHev9lrARfz3Z2viHE6bu2djSX9zHTuNAagxHLx7tUY7LyWK
+JhOF5a/wb3GxrHAsmxWrA7gh11lWFzfMdDBcaTiBY8jcFu+tmIwc1482oAOpbOVU
+Lh5vKQJH7EDwU8jkla6RkC/JO/ZZ9CUrJkjlHbvoMFuBERMYBPD5CIM/hd+ADiwo
+zhKvFcX/MXhF7RmckMxJOKsTPWDQpdFHWjUTefr00SLVmiNbgl/fVJZ/by+/ZSjd
+3tfSuvgyTphpumfQRi3yv0UL+POiVInyJ14qcIAOAVY9/svuIZ2T9WIbsA6OIzU4
+bPOIeIlSOktuLlXCwUPW7J6/WJ9yxCAhrkfFiVFgYm5E/W2WvGSvKC2S2wW/xnPQ
+wyf3bSxHqyNKf1Sd6XqcyC7LUw+4aSeTue4WyVcYZEG1D2haG3RNRRKdYGXRduGP
+3cPQSQjL5feJLcoTPus8vg0P71o7SeSwhyWqTaTVQ4dLO8TDdKc0XXR2ERANz2re
+hNg+ziqAt0vA7BXC/dEMtlGj8N9eg71LUp/uTyT1W3t3FBe8vaXdbmCb/mD5hWWm
+i9VUinz60iW2bRB2SvzeQFfgIThT8XyfJCkg/d0PGAv89ll1a3zX04ajvApG+6hC
+Bo7xG2p4OQ091EUaPbqdtJUbsObAhXIam8y/BBzKzCJAXjZ1lfkOFvOdW4dKCNsT
+68OzOGzIN9bn+8sRgYB9GWhtmnXYwtsbv43pNG6wEMiaSgn/B1pdgdu6GJC4BYzk
+j/+Dpq2I9As76cFV6m5kyMm1D9/GVnR/z0V++aoMM3kPH0ZAEFJun7iobW6k4u5z
+c29MfcbOzBM9IMdClC0qHptm6IHAjdQfQeCG6x9U1zLIf/fQx5OsiAUO21LgJyjF
+s/i/9dbtwpqybPISO+HQxznmW09cWt2m1ReDRbBLDpIGR9VaSa8PGl8dVxmutTwq
+QVXi128+wMuBpKQLazDqjSWGlr5DcmUrHV5PzyxoWtFPyJrsSr5g6UsueS+qOas1
+H71dvlCwlEc61q3LKbVFXsTN15qZs7786jgSLp5oyXCLkE6gKARs4zWHRE5zNiBp
+t/kCbo/yuL+vpUbW77rqknHx/Sc53pOb8tUAycUEZVJDfRVx15P9e1ZZuVq6RnIg
+8/f0YabL7HFCkg/zh3zZPYoparOXl1eZVDlXZQa2q4H7Z5218YAY5GwyYDti2g7P
+ZAhZs1dG0pdnuukEg23XmmgTh4JwKi2amIMZ6cS9vosHQKF3T1lVpp/JQoMEqa8s
+M1n3qabVghV/QMrtwKg9Pau2AwOc92XEtx+TpBtdZd66Rk7m63Q90JIYGKk1rZ2g
+SMOyju3OrVF8JcpULTp475LBHl2Mpmy8I2m1CjDA65b7cV0EXg3I+v2e/Yswsdwu
+dnD88uIMGxMG0ziVfdPH5Y5XCnywWANPXxP71GJhBqDVljOlg+BLym158u0ueqwH
+O2z7Idny3KVKBR2BpP3Piz0DlA3gXftyQpkBv8JxMLfbd55wOT5DWZUzDcqaccbh
+ofv/CUameNuXHwG5e4Oe1um38aJ5yW7dYVdMN0bdazXSy2O1ADqtJQHZYRBVYVzX
+xFkKgYWgKpVShOBSe3yWD3GpaGwU/tE65SS5Yz1eXDmNM2i0+RvuT4K0LlB5VGVz
+dCAoRmFrZSBLZXkgZm9yIFB5VGVzdCkgPGZha2VAcHl0ZXN0LmVnYT6JAk4EEwEI
+ADgWIQQGvdIIuy1wmsNUK1flLDyrx4iGlQUCWr/9iAIbLwULCQgHAgYVCgkICwIE
+FgIDAQIeAQIXgAAKCRDlLDyrx4iGlb4jD/4ndxhZiZ0z9xGgWSMXecPVhY8eQ21W
+JU/fQpa1AT1IR2ZWZyFshcZa99a83FEQdivZzaRVmqBw1ht9hD9fgQ23mMWuZGD0
+BvWBxV816wNnd9gMaLTvSJMsbddFiqd3yWNfzSE2Z6b2HHjkM5aMhLo2XieI4JwQ
+O0xHVZCKNY7dsnJ0rJO+oP+a4B9w6+yQDPY01hQdwBGdvtI5K4zHuR34tYbyL3Ve
+xCSVeKk4IcycQzPThQuQd6g6P3MAnh6S+Zlo1bn/2sHs+6+IHDB53p4kgOoj3BtA
+qpDBGtV9rt8UrmzFhkICkMqvi6WNztbZA/ViG2DpJrY2nDL6qjGrsNescYuc5VN0
+5CG6TTNapeuBb8Ksck5N2p4V219BOrKneg0W8imHKObQMxRC3KFEyhZ0STvHKIBY
+CZ7myYYnRARVtC6uaSBWhZPjxM2yFhCfexQp7wevvOvYirBGCzN+nDPrhQ86iJ20
+VAL3GwHarvTASioujQefzx/BcHMuOL75TAQUgc53s7MwAm4n2SmHmZV2TlU6EvxT
+Q8G13SbwNoxolU5S2bu2aIzeVaxYAppJH7kUQ6LnXOyBKPbEyAvJv9OY1d4IsJ7d
+czh2/3QQe2E2ZDd+bZMlPqDt5ffnfFc1vqQlPcYB13vwa5rQvEir3q9c61scvPz8
+zAJWx6n4AEdVOw==
+=l9D8
+-----END PGP PRIVATE KEY BLOCK-----'''
+
+PGP_PUBKEY_BIN = 'c6c14d045abffd88011000b0d91d197f360568ed00f70dc0d083f71f47de77ae43622b5c0906b60f78ae0df207366125b23104c5e8acba6a738534a4a3c869094c7a93d738727696fa27b109a20a6556a78e94741d2b842c50d79c4bf38f0cac48b23c932693b628e5bdaf4d333d223d7639d193ec3317ec69dcc8a30ba2dcc1dfff31383a21c50f98a82e85b8d949ac492f6c4b61daa7206c88175ff2f6c23b785167e2a20cc847cf5adfa0600543ba9177283c8c3646e89a7e59f8bd5be409cced4874a9a28656c85ed458ea937d835b112b516e43b724aa48e9243b3928ef287522dfc3825c61b762fbf19d3d1a06fafdd1f3da262b9e1a0d46b800f72d3b31c3f88afb59250aa88040bb3a0508260e95bcf03430b087c8335e9c18b19e67c67169a8942a67d2f8aa849edf76a0cb6a4107d8867161abeea1822b1072c6808f5c0f24c0a7c00c53a7a8c3b19ee7c33af7839cf82f9471d085c684f12907e7d374c99a993f1bcbbd63edf7a69865f153d3acae145ac5747989f1324ceffa204fbd94ae1f9392f31ceaf941aab1024a4916b0ed6c4006c61064f906da904f535859e990a28d74eca89a86a512e187c048cba8400b9f3551de229b40ecec73d9ab3dd5cd573d1a4739150acb9975ccc9915e07601e9527c0846d233b52ed937bb99cafb32acf74c228c087c68bc824443a701c809112721953e1cb57d1d627ab6410ca03208dc574d5dedd0011010001b42e507954657374202846616b65204b657920666f722050795465737429203c66616b65407079746573742e6567613e89024e04130108003816210406bdd208bb2d709ac3542b57e52c3cabc788869505025abffd88021b2f050b0908070206150a09080b020416020301021e01021780000a0910e52c3cabc7888695be230ffe27771859899d33f711a059231779c3d5858f1e436d56254fdf4296b5013d4847665667216c85c65af7d6bcdc5110762bd9cda4559aa070d61b7d843f5f810db798c5ae6460f406f581c55f35eb036777d80c68b4ef48932c6dd7458aa777c9635fcd213667a6f61c78e433968c84ba365e2788e09c103b4c4755908a358eddb27274ac93bea0ff9ae01f70ebec900cf634d6141dc0119dbed2392b8cc7b91df8b586f22f755ec4249578a93821cc9c4333d3850b9077a83a3f73009e1e92f99968d5b9ffdac1ecfbaf881c3079de9e2480ea23dc1b40aa90c11ad57daedf14ae6cc586420290caaf8ba58dced6d903f5621b60e926b6369c32faaa31abb0d7ac718b9ce55374e421ba4d335aa5eb816fc2ac724e4dda9e15db5f413ab2a77a0d16f2298728e6d0331442dca144ca1674493bc7288058099ee6c98627440455b42eae6920568593e3c4cdb216109f7b1429ef07afbcebd88ab0460b337e9c33eb850f3a889db45402f71b01daaef4c04a2a2e8d079fcf1fc170732e38bef94c041481ce77b3b330026e27d929879995764e553a12fc5343c1b5dd26f0368c68954e52d9bbb6688cde55ac58029a491fb91443a2e75cec8128f6c4c80bc9bfd398d5de08b09edd733876ff74107b613664377e6d93253ea0ede5f7e77c5735bea4253dc601d77bf06b9ad0bc48abdeaf5ceb5b1cbcfcfccc0256c7a9f80047553b'
+PGP_PRIVKEY_BIN = '950746045abffd88011000b0d91d197f360568ed00f70dc0d083f71f47de77ae43622b5c0906b60f78ae0df207366125b23104c5e8acba6a738534a4a3c869094c7a93d738727696fa27b109a20a6556a78e94741d2b842c50d79c4bf38f0cac48b23c932693b628e5bdaf4d333d223d7639d193ec3317ec69dcc8a30ba2dcc1dfff31383a21c50f98a82e85b8d949ac492f6c4b61daa7206c88175ff2f6c23b785167e2a20cc847cf5adfa0600543ba9177283c8c3646e89a7e59f8bd5be409cced4874a9a28656c85ed458ea937d835b112b516e43b724aa48e9243b3928ef287522dfc3825c61b762fbf19d3d1a06fafdd1f3da262b9e1a0d46b800f72d3b31c3f88afb59250aa88040bb3a0508260e95bcf03430b087c8335e9c18b19e67c67169a8942a67d2f8aa849edf76a0cb6a4107d8867161abeea1822b1072c6808f5c0f24c0a7c00c53a7a8c3b19ee7c33af7839cf82f9471d085c684f12907e7d374c99a993f1bcbbd63edf7a69865f153d3acae145ac5747989f1324ceffa204fbd94ae1f9392f31ceaf941aab1024a4916b0ed6c4006c61064f906da904f535859e990a28d74eca89a86a512e187c048cba8400b9f3551de229b40ecec73d9ab3dd5cd573d1a4739150acb9975ccc9915e07601e9527c0846d233b52ed937bb99cafb32acf74c228c087c68bc824443a701c809112721953e1cb57d1d627ab6410ca03208dc574d5dedd0011010001fe090308feae6e82eda151daf19f4e69be97e69011fe7f387fc46c2cd5c08a4f7b8cc455e5462e8df1da7fa31ba53d156206dcb3689b5f848e2fbd79d2562ec629630c0bc1e7c4efbf42b4ae61e8a496bc489ef5e1ea3d81500528413d7d431e830acae24aacd5c8cf50dad7070af71c54345034a68d7abe3723daf9f57624287629023c050a3611d0ac561249ec1bcff7dff2c69e406c63445abf5130d743ba90c7ad153e386ce9cc4aa40069d02dcd164853962412677feeb79435ff27c6e6299e76d4f90d9a55f7fd33a13a594857bc959f90aced6e6f7a633ca606a779d88d40105bd5ab018413c8f4f80fc88e6e346d5d403d1f5ec893ac70ae14688eab8b0eeac880004ad1521692d2826b01f05cc9c456be94f7df58f13bdbcb3487d5233b4012110821e74e3bd31c9fe27e775729957e60e2db4d187d42639c3b3021282c3a7dd88d78c61806cb5e7aaeadbd9a6dcf06589ac2aee6f2d3bd4cda6dabb7825a0b5cf3936e758cbcf1e5f5952b2fc8c0a178c6e1db819fe79ec83171d320de49450a10ab4956ed084d68406b680b120aeabfc6b097851cc4047be04b7b0f52ace0d62c33302a148bd0371b136e0f7e8e25b0aca2f8e9e68d3d851e45963d5e6795b1aabef9a6405039a367dc2c0c4850e3d4ea4960b7e372c78378803e419120db1d45f75f707764c81e02dad3114f83d64fda5263aae5832e5ac7d927ad4a5234d4284eb09ae8ca8f580b08494488e4910f10c92ef57221aa955781139937312aadeade7db6fc28287935e0e6a1c3c251db5168c603077ecf5c6d80a630b8ccce6e932c99804a5b9688dc4448aaff6cbcc730e67e4fc5f9ac7e2c4e51f24813fc0b9a23228713a8af9b91097daffdad5e3d25b95f2130acc6ed139fd7f1acc1494c4db6e12f302945b654c793d3da9d0321d61812b7af6ebcb0e794fea36419447f8b0606137ec619f309a23673a93e6be996852212eef444a76ee04fe522fea506a7870d891e7c63f5da5aec4a4be232f0182a0b4b7a33c000519456f3e035567e1435eb1dbab405612d98cd5b3adae1fdc5192d375b7741fbc7bce763a97ec29b702a56f3edad1db71b66e6c71a79439c9aa8182d9ba008cabc14b73d5585f380f47ebab831e16c9aefe0d73ee6631002d6f40d80f1f7c086df512a93199b6306ad681687856f6df8df06e5d7527a4063bf85db1e32c093443cafeea9889b231c8a820e0d884e28306b21c6090ca7c9e2fea4fed6462628a9ea20682b958be51624cc4eefd0b75fdfe547aa508ef33e624d45d72b1a9e9fc877e4096c3a8d5bdfa93c9cc57dad0aca89069d421fd87ba1c11b346f327ddf308cb0c876f84b5b4fec74c3135786fde73a8755ea9df3bd038797922f3b98df124033a7c8cb6372636bf249e04c65547527f693dbfd4b1aa5831c1281e9b9f83cea22b5a065b9f07c7f77906b58d78fcf6eb681526ffad6edcf93cfd1c83724ce403df825a68dfbeeb9a31bcbb65f69980f6026bfb6fb28b19d26286787274232c234a10d404ae83d5c88d8a96af264a89ca2b8d62777368cd81c4d2c06d26b63f1b9a91b3d7498de29daef0f617cce31c6199bd037d330824e83709f00c381a60f629c1b2bcae5f2b3068e32028e19c38ee4f2408836b4eaae218b19f94d581d83ba3e0d4d769b3e9f3db27607c767fcc34abd29e1b954a46a2deb0035a6967edfaaae9f76d56d7d8d466dc0c0ef28d0e90f3ca8a19b53089e40900cfb6097c49ab33ce9f88253fd985e2074d96a256c17f982772f1d3760ec85063e6bf3cb746cbbfdc8e28a1cc04f1ae6d65df61361e4ad1cd2ac1b3778abafde2eef60932ee3c349f5e6dcc9542638079e3b1e5786f4a9f9b616cbcd5806ea8f52b42e507954657374202846616b65204b657920666f722050795465737429203c66616b65407079746573742e6567613e89024e04130108003816210406bdd208bb2d709ac3542b57e52c3cabc788869505025abffd88021b2f050b0908070206150a09080b020416020301021e01021780000a0910e52c3cabc7888695be230ffe27771859899d33f711a059231779c3d5858f1e436d56254fdf4296b5013d4847665667216c85c65af7d6bcdc5110762bd9cda4559aa070d61b7d843f5f810db798c5ae6460f406f581c55f35eb036777d80c68b4ef48932c6dd7458aa777c9635fcd213667a6f61c78e433968c84ba365e2788e09c103b4c4755908a358eddb27274ac93bea0ff9ae01f70ebec900cf634d6141dc0119dbed2392b8cc7b91df8b586f22f755ec4249578a93821cc9c4333d3850b9077a83a3f73009e1e92f99968d5b9ffdac1ecfbaf881c3079de9e2480ea23dc1b40aa90c11ad57daedf14ae6cc586420290caaf8ba58dced6d903f5621b60e926b6369c32faaa31abb0d7ac718b9ce55374e421ba4d335aa5eb816fc2ac724e4dda9e15db5f413ab2a77a0d16f2298728e6d0331442dca144ca1674493bc7288058099ee6c98627440455b42eae6920568593e3c4cdb216109f7b1429ef07afbcebd88ab0460b337e9c33eb850f3a889db45402f71b01daaef4c04a2a2e8d079fcf1fc170732e38bef94c041481ce77b3b330026e27d929879995764e553a12fc5343c1b5dd26f0368c68954e52d9bbb6688cde55ac58029a491fb91443a2e75cec8128f6c4c80bc9bfd398d5de08b09edd733876ff74107b613664377e6d93253ea0ede5f7e77c5735bea4253dc601d77bf06b9ad0bc48abdeaf5ceb5b1cbcfcfccc0256c7a9f80047553b'
+
+ORG_FILE = b'Hello PyTest\n'
+ENC_FILE = '6372797074346768010000009b020000c1c14c03e52c3cabc7888695010ffe247128bb961acd49999271747cd81ee1a116a05dba3ea998d6659b5653547eb97c6c5d8e061c200c35757dfea7aa18ac69bb1acb663a1d2c0cd4b5cce20a3ac0914636d8ca6e142b5a8f19956bddae7f6646f068d4a6c037399b6c7ce9cf104f728cc74ab96072bc46c5a2f575920f0dc89c06f0a9d411d4f2e26ed1341bcfefeb1a155b0ef351bf469bb0cea14779d9ce01f2cf77f41b186c09960a2acdec65debb0417406168277d2a052501b303053e536e57aa16f780521876c40217e31e368a2e5b7abc2fb05fd6d30efbf7ea70635b45b609d3dc3a8010f431eb70cf4f55480c2644097edafc4f225e6638b86d142b7af690667890873538b12273399effa6dfbcb1c59b77539f332a259dac615445b778f77c8aea5965d13b01889cd17492b6976ef987901212d2fbcf93aaef005ca40dbfdd9f3fc26fd3331e1a79aa9fb1eeb15533adae12e2101db9caeb1697c485f1b26bd406f3f51473a5a9afec0fee7c6b587c1dc5d7b960114a459f50f88fad2c4a4dd235eb96a5456ba996f992300e89ae3a28345a3c045027b9145ebc4f392d327718e342510a4412337b7ea8e5c12d0517bff0b91f16bbc3a6ca8ba360af06e743ffd61dc8997beac5eab298a092c8f6a4578cd0fab741eed17c05fd92de4615b25eaeba6dfdb5aa423492d694293d7218845b94c81386a5b5af2bc958a1e52e1c540922f4694d9e2fa92cd27a019c472abfd95360d42d1d057c1d223951c3614893e03bfd2312bce7546d2ac7d77f41138242259e710f8c86cc7d810d0f6f022021fa61c39aa0e7b57e6575ef3193bb13347cff8ed74bb1cfc956dda6a16fac59d337af6877eb10b94256b2543763d0a1577c161f14cdc319317e9045821300c2b108e86aa322e426a1fe5197eb5c0d186645f3699a9c3ee97c0cb8b1f0a41a45a79603c1ce25acdc327410674a50f38298505e'
+
+BAD_HEADER_FILE = '6372797074346768020000009b020000c1c14c03e52c3cabc7888695010ffe247128bb961acd49999271747cd81ee1a116a05dba3ea998d6659b5653547eb97c6c5d8e061c200c35757dfea7aa18ac69bb1acb663a1d2c0cd4b5cce20a3ac0914636d8ca6e142b5a8f19956bddae7f6646f068d4a6c037399b6c7ce9cf104f728cc74ab96072bc46c5a2f575920f0dc89c06f0a9d411d4f2e26ed1341bcfefeb1a155b0ef351bf469bb0cea14779d9ce01f2cf77f41b186c09960a2acdec65debb0417406168277d2a052501b303053e536e57aa16f780521876c40217e31e368a2e5b7abc2fb05fd6d30efbf7ea70635b45b609d3dc3a8010f431eb70cf4f55480c2644097edafc4f225e6638b86d142b7af690667890873538b12273399effa6dfbcb1c59b77539f332a259dac615445b778f77c8aea5965d13b01889cd17492b6976ef987901212d2fbcf93aaef005ca40dbfdd9f3fc26fd3331e1a79aa9fb1eeb15533adae12e2101db9caeb1697c485f1b26bd406f3f51473a5a9afec0fee7c6b587c1dc5d7b960114a459f50f88fad2c4a4dd235eb96a5456ba996f992300e89ae3a28345a3c045027b9145ebc4f392d327718e342510a4412337b7ea8e5c12d0517bff0b91f16bbc3a6ca8ba360af06e743ffd61dc8997beac5eab298a092c8f6a4578cd0fab741eed17c05fd92de4615b25eaeba6dfdb5aa423492d694293d7218845b94c81386a5b5af2bc958a1e52e1c540922f4694d9e2fa92cd27a019c472abfd95360d42d1d057c1d223951c3614893e03bfd2312bce7546d2ac7d77f41138242259e710f8c86cc7d810d0f6f022021fa61c39aa0e7b57e6575ef3193bb13347cff8ed74bb1cfc956dda6a16fac59d337af6877eb10b94256b2543763d0a1577c161f14cdc319317e9045821300c2b108e86aa322e426a1fe5197eb5c0d186645f3699a9c3ee97c0cb8b1f0a41a45a79603c1ce25acdc327410674a50f38298505e'
+
+BAD_ENC_FILE = '85020C03E52C3CABC788869501100082E9DF4C2A9618F1FCC93BED449011D0DEF17711EB39811857F4968D9A2CAC8A07EA76A267C4B9CBB8CD9660AC8BE22CF8A57CD037EB0D21A351BC884E5D98AF85FA45F81DC96A21B153610D90B4507AA54A9AE909548C7ED9526C0C4B0D2431E6A05E8E149BD00BE350F3683C0977D376269B720E08A6AF7A03D69F320B55CB2C43F3D4FCA3F27C7490AD3840F02DF57A7289320DE45AC6C7489C93FB9C56BD05940209BCE4BE2E969A09BEF6543E7D30BA852BE4396FC48E6741076A809D5E5563DC2585E0AE6C6D0E5D35668F327AF8F115AB97370C7859E9E39BF86281FA03080410C96E25BE20CF81A91291C149704395F635AEF9DE324A207099C4B1CA8360EA2B65CC6BCEEEAD3095EE15A7D71BB5C005951B15D6A41305F0124CC753B3C6878FB10F9DC49261394B0B3999B32322C66522EC44260AA0BDD2F8E62CC6CAE24908EACA3C181EAA643E03EBA37B03F76C3E0023BE37A09DED37E85EDD6512CF73548EFDFC81199CE9200CB88C1A2FB401E939025E6449C70455F415CE05F585C8ECFC5A7FD90D52756CBD6D220755F3B1E509052A1CCF0589D0111EC37BF1FECCA5D91891FB216F921655C49CA29BD73C14FBE51E3608F124288974CCB1D4306E83E2CC659D5DA9AF60E6E7A0CA142240CF664A6C3877D15CBA6F1EF57B3D4A9F51A47141D59897BB85ABCF179E585077D44DF20E7AFA0415D61C0BAB0FD25201384278D02EA162773B43DFEC4B56C223E1477964AB0222A3BF108F703D3DB62BBB534654349DF3A51E35856E48C116E280DCFC193B8E1E28A4A349B9CC7847BBC0C16AE817D3333AFDFE027938328232EB'
+
+RECORD_HEADER = '<Record 0|18446744073709551615|32|0|0>'
+RECORD_HEADER_REPR = "<b'crypt4gh'(1 | 1 records)>"
+
+KEY_ID = 'E52C3CABC7888695'
+KEY_CREATED = '2018-03-31'
+SESSION_KEY = 'b029fe21569d68018f04ec7b98caf98b54496180b6ed5805a7f8732b841386cb'
