--- tests/rsa_data.py
@@ -1,73 +1,0 @@
-KEY_ID = 'rsa.key.1'
-
-RSA_PRIVKEY_MATERIAL = {'public': {'n': 423207781184939113316689143913078115613226662612618524081225039828382937601309048028362822644903017813934802715017389724535033876145349059985521062585021115561842672819226870692196458273564886931327793507599992449310646681067994336149674703436807171770303174117405390262116427937704047922214607092466687546608637379852067147926522411500275085850638185354987984094201206419551225865952837041312615724687739057366102580012744748173960347654827824301724522899184771218511979521678541813453166775083294024019388671429556854516580873757532096031127929627394874141320688933223639193381922205298041723006435486089417651041090938485695791381435037470059427013492830004817202303203613844425719539492241307801756196466737565060589287733566152271985183197735464457669137772958433482486572270195138066513436400854241441979662307329450721697171103908153383071790900383071764216513253803420715639561290934856785513837766026983386177250377580809885006799459001407233680142778528927068312304464499217992762947013704410676931867522919608743553830461366124871891351062638834183723753110249418850670478433597840827647638495445145395965316350673628573362445933201889069083548805742833465648187096844303274700226238720073733546958253827324437360203620001, 'e': 65537}, 'private': {'d': 417292674793657787839506983365229121092548028818682906520787391835908476291777041863997954775718722464474737724447849938037725622326145560604305302124092489871703547007816342081578762074798000494214464260717138594487133972829010513089218746826829367334006155524983653876866894880180253609128128002842513632747863895561597741217324637282134921078994921583604353573574868548115106988170625485552627382166507249737627825854152286063589905332905516489948279510325756456832974177493462479624595116844798253294427809274904153313639205991813503480896561338661888122530543655550983235661278282932767050405097388445981018782116991474720950999241933188587234697360925823283459682573005635200795924295668078007458339778134258978219078968010268576565561857799917061670971549442038348884326673815281573484595596530511848385703684340140662836029047845680613012468545558260178246762123983516611203753449044180358048809068987027730640782540144643511990482015960741964589277694647805762903384206996531275235119651880774555954773406916463478410907836999134151609940571422264482973983822991307481426314943030564481009870784698060630603613961769601938293392995757571151739615135021716611510458319226564949960636612614835489000631618973876206117483997829, 'p': 26174715560556250055781956315888694847156411868120469726117105893318057623292273085594069988627111721991417684716119083762171974143441300025159779238896347909047969326969475297367053079562058723031904549698364500420152807860382108125815037803787433884081935944582084403257160962021573114040887941533095119268007561253171319268556831983058766748448420559794885315154161557366854983865852141416813697603275462315596717874677416868054356619611107834728134872425457192710397742272304365133849239059915098154990414931241688844634036066902395962812689237268043542831865913607333245001729178313240583303354122643928560129003, 'q': 16168572308105163763177530109091306669025201860610636547906198086500575188996715371447603118218391460020568887826697558901626109238590612949057419358526514642331140412092261826862408701929735190926338147515180113684522096372134140215261445116365050250389126451882174641022430702177545514738057399099441596835367348044637507087582980008621291037986090664079866484206298598391340685456541512045779367027475681005868615848541048529112536332883535791131445295730051681194129209213555207979312076981945671651414382056576656902443330042148248391967830863006094749542409079580466898575302922881307684208876957152829957525667}, 'id': 'rsa.key.1'}
-
-
-RSA_PUBKEY = '''\
------BEGIN PUBLIC KEY-----
-MIICITANBgkqhkiG9w0BAQEFAAOCAg4AMIICCQKCAgBnvIfMoNlwfdTQckG0/moI
-K/MIdRNMUjli4B6A5hSBzs/eImLHjfFa3CJdfdSdc/Bzw8NAMD/32xSe1XC6GiU5
-bGRGk22os8R7gOMraXlfuL7H8jS6N7TQsoUSZiQE+L0zO0jERC2+Yf9DPV7ws5tm
-7z9ZNFRA0kWAxG502AL1xPnkK8iniYgUDhtOSR27/XTFw5b2w/Rj+TUHa8YGrbsl
-9rrS/M4AN1vIKntWA2oE7iXT191NyGSV71TJRbKFebUiewW5sIjYeOALGiYfwXVj
-EUumU36EE+9ZNA7y60EkifDYVVBXV+RirCD3lk5gPbvHqTHFdOyKI0m9DbIotgUa
-KkBS+naA5JPMwH7yU1+WHu0oeTrpjhuS4qKKLieptfe4RAy5Ltsus1uWCMehQrh3
-u/eKrjRBwilHtdIVWfVzy6nTWImnEkbFGp4JBfuX0OqAa83KoM3HoED1yl29cWqF
-/gcZHj18kLt2H+uQ0RvSVOkdKJuhFg43rir+yHgvscJqDWVYsgj7mQvrPxuIjd+V
-+toBQH7AGTpCZzZMOV2/zVx9/knmNLDvI18eJHgSQQW3NKR3/EDKW1mBa7BZwBL+
-WuT6UOT58pRjAuwfum0xgqSG7hZ9oIep1GbfcUzdx+tJ8hdp+RHlIc3gJC87PIcr
-y/0VwuKLWyWiXLW0uPleoQIDAQAB
------END PUBLIC KEY-----'''
-
-RSA_PRIVKEY = '''\
------BEGIN RSA PRIVATE KEY-----
-MIIJJgIBAAKCAgBnvIfMoNlwfdTQckG0/moIK/MIdRNMUjli4B6A5hSBzs/eImLH
-jfFa3CJdfdSdc/Bzw8NAMD/32xSe1XC6GiU5bGRGk22os8R7gOMraXlfuL7H8jS6
-N7TQsoUSZiQE+L0zO0jERC2+Yf9DPV7ws5tm7z9ZNFRA0kWAxG502AL1xPnkK8in
-iYgUDhtOSR27/XTFw5b2w/Rj+TUHa8YGrbsl9rrS/M4AN1vIKntWA2oE7iXT191N
-yGSV71TJRbKFebUiewW5sIjYeOALGiYfwXVjEUumU36EE+9ZNA7y60EkifDYVVBX
-V+RirCD3lk5gPbvHqTHFdOyKI0m9DbIotgUaKkBS+naA5JPMwH7yU1+WHu0oeTrp
-jhuS4qKKLieptfe4RAy5Ltsus1uWCMehQrh3u/eKrjRBwilHtdIVWfVzy6nTWImn
-EkbFGp4JBfuX0OqAa83KoM3HoED1yl29cWqF/gcZHj18kLt2H+uQ0RvSVOkdKJuh
-Fg43rir+yHgvscJqDWVYsgj7mQvrPxuIjd+V+toBQH7AGTpCZzZMOV2/zVx9/knm
-NLDvI18eJHgSQQW3NKR3/EDKW1mBa7BZwBL+WuT6UOT58pRjAuwfum0xgqSG7hZ9
-oIep1GbfcUzdx+tJ8hdp+RHlIc3gJC87PIcry/0VwuKLWyWiXLW0uPleoQIDAQAB
-AoICAGZJWqnlzKGEIYyrrKh32DtYgppLPo0Z4mXhUZKNy5i3QPXqd/YGuv55YxVG
-lpfukyDC3cp0CQ8vIWUAVtfaIew+OZ3xjsYzA8E3tEfBB6Hf+5YR9/XKKetCq74Z
-otiYH47dnQGRrH02CcnWQMAciVX1AETIOW91Jn0ysytZN4n5AOIAXij8joYf5sLJ
-i50q7LJ4Cmoli9vsBWLc3PUG5ysm3s5d4g4Kh+CWKukEEszrubbp+dfRiYTh6XtQ
-0xeTi4yFO37+V+EDRv8LeJ4FVucpTauiVnSLZ41QrJDQcXpC3Y2iical7pdw3Ezl
-E0EIfhkAdwgZI4eOBg5467wCxoHce+vGWXsxxs2yWHOJR0/gKfVKV8Z7XA6AoVoj
-JlBoxCw0JGQkiH8gBeOF9hfr+WNSkOZX+PUKVnQx+kyD0/RaXO5PbAzimY8ZEY8Y
-QfqoSzi/yycOHTgpQi81Zzsj9ojRdN0IZeJ6xBTiW4LOMpRqASFxm/J2micfyj93
-G16MwdxUpDNZzlGesEOq36/tlYsc2f6NUUyeypPWa5A+3uA6MtwCWzSUgXxxXU5X
-Mr0X+125B5/58VcNE9MK024HDqy1aN5PCk2syngg7UJWxo2wGAEqUU3aNsnj1gv3
-HqMTSXiAZB9VGLrZhZX2sUYTsJgK5hG7W68ja71AD3sK/GqFAoIBAQDPV/va0Upa
-ljCgmDGR/XbRAOoh5HQPE0+vljaSfA2sdu82Hn9+0Y/26EOZRxWF0y3AHnmVBseZ
-+mcuW2nfi9EzirnTlyNNgpA2dRXPTU3SlFvcGLCYhgQfgz4OCtZ67Uqsmb6VrsZU
-gQqTwSW/6D+muTWXOqgq/dO0T9fi9S3NcgsPnm/0jTT6o+3AJt6A1Vqvd4upfQtC
-4NLYVWPhzSZXyi9Vy7Mu11D2T85eVqPFCc3qq3zsdkvPuoB7p89Y8hvZVAt5GFOl
-jFexFbccyTC4+akda7tSjH1hcNqAnlYEJIQvOdKxLsxmAYS6sELCPpi/BDgsoRBb
-nHgPw1AyV1PrAoIBAQCAFGtp5Vg3GVbC1PRBuwPfZmBq1RI2B8yuwaHd4dA26oUA
-YTJV1ZMOLm6pil740y31uvqmEL1uVyPn2a7hcm7b5Am/OG9fNsq/JgGHBmHxqvvB
-2Wr8W+EgJBo2t+2BrxGqkEPTIVVqhlU58ibLePYT3sw56fZ6zIOlGtxZtoBI8aJK
-0sIZXYMFy6RREk81EtlmFszDRQ+jzpaoopDabTTb7hJwCTHjFEXfmLEYeauTnu9D
-8BAutrPPlgzQ1pHxmBbNZmccwoReqOhcfhq7Ot6NqLDHYBQ2N5hN1WF7R02dAadt
-lilAZkWjCm7MgqHLfy17c6G/9R2Ocxe8zHHJkNCjAoIBAEBZ2jzLf/wlaOou0hOx
-StMLKZ9jO/IB2FVylXVZMxU3xH5TcNZ3S+shovlLao6iQCzESI/MDt7nLGbi0DUl
-O8R3Zb0Zx0cjIocER8STBXg5u1jQAN8VmOn+3Z+WFYMJaoevH8ts9vj7I8E/zOuG
-tVqVWOCav9638KChtyslVRXSM0Shs7V+rdasUfm1FpF9piuXOq1btccH5+hkeB94
-gFH22kISeoYrJWGM20vEMAfTRylyF89714OdKos2RdAhNCt7PiRTPILAz+aSCsbQ
-WmLOeYI8EWNJXxCC6g3JFoEqpCvUEmQTVjTCe+NFGTMYImgGczjGtED6ivi5cYZu
-PAkCggEARFWgsZExrcQRLEYeQ+uZt1mR263u13D13c9lo1JVmtcpOviguHcIzKot
-VLuEzFEv1kl6bZvE1dgcShk0WwiI6T0Tjjh4A17icXW/nuDi5OOLCvQLRrVmuZVJ
-fyRGzcgwLsZlyYGhZwJg+leeJzVUZWmboNKZNQ9ktXRNWHgsnb5+DlnAQXRI8ZCj
-DLYuGG0EHJnRieWQ4J7adwK1MjTN3QF6SNg/C57HYPLAmgRgbJSBp8evepzz9q9i
-CxJQL4wrrk8SeJ+3rCGGR8LoVDimgQIdv2Idpoem9PAB4Jo8aKZ3Q9q4C6oFQWXR
-k7+mv0INbC5XgGkw6c2yPoL9usNO7QKCAQBMeBBflBLFcGFSyL30CgyxG3UAzY+4
-eDDfzLOhSwC8Zg98d4P04I9LLHH75E3b0Fh5TRrrKAmnWLOdis9rKgOulX9I77Ik
-uDTu4E4cgWmyD79smnQSYFJSp1itV9e9+yJ08QXcbSN8tH15lCN7ssQXaAR2Moya
-26otWMs0yHSrRARQ4SjygiEzhoxJghX4DejUBj4PZWNqtN4Q+anI4LXCjrlIEws0
-MsiH+C/s8tn8UEvOM8sCibX+To3YhotQ3qPtTYOKYswBQZrcvSM0pG7rRYBEdvnE
-4eXcpE5IOw4qNlRafd5QJtsMM91my3AKTOuopYUyN830dAavNKeNVlB2
------END RSA PRIVATE KEY-----'''
